split TestFactory
