+++ shared/src/test/java/passoff/chess/piece/QueenMoveTests.java
@@ -3,7 +3,7 @@ package passoff.chess.piece;
 import chess.ChessPosition;
 import org.junit.jupiter.api.Test;
 
-import static passoff.TestFactory.validateMoves;
+import static passoff.chess.TestUtilities.validateMoves;
 
 public class QueenMoveTests {
     @Test
