+++ starter-code/1-chess-game/passoff/chess/extracredit/CastlingTests.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static passoff.TestFactory.loadBoard;
+import static passoff.chess.TestUtilities.*;
 
 /**
  * Tests if the ChessGame implementation can handle Castling moves
