+++ starter-code/1-chess-game/passoff/chess/game/MakeMoveTests.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
-import static passoff.TestFactory.loadBoard;
+import static passoff.chess.TestUtilities.loadBoard;
 
 public class MakeMoveTests {
