+++ starter-code/1-chess-game/passoff/chess/game/ValidMovesTests.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static passoff.TestFactory.*;
+import static passoff.chess.TestUtilities.*;
 
 public class ValidMovesTests {
     @Test
